+++ database/migrations/develop/2020_06_23_233532_create_meeting_table.php
@@ -21,7 +21,7 @@ class CreateMeetingTable extends Migration
             $table->timestamp('datetime')->nullable(true);
             $table->string('place');
             $table->enum('type',['ORDINARY','EXTRAORDINARY']);
-            $table->integer('hours');
+            $table->float('hours');
             $table->timestamps();
         });
     }
